requirements.txt
openai==0.28
discord
python-dotenv
boto3
